requirements.txt
pyglet~=1.5.26
numpy~=1.21.2
gym~=0.20.0
shapely~=1.8.0
matplotlib~=3.5.1
descartes~=1.1.0
box2d~=2.3.10
stable-baselines3==1.2.0
pytorch==1.10.1